[CELEBORN-20][REFACTOR] Extractor CommitManager from LifecycleManager
